ДЗ Урок 6: готово.
