+++ lesson6/lev1les6t3/students_6.csv
@@ -1,9 +1,9 @@
-Фамилия1;Имя1;ВУЗ1;РТФ;5;РТ;11;Йола;18
-Фамилия2;Имя2;ВУЗ1;РТФ;6;РТ;12;Йола;19
-Фамилия3;Имя3;ВУЗ1;РТФ;5;РТ;13;Йола;20
-Фамилия4;Имя4;ВУЗ1;РТФ;6;РТ;14;Йола;21
-Фамилия5;Имя5;ВУЗ1;РТФ;6;РТ;15;Йола;22
-Фамилия6;Имя6;ВУЗ1;РТФ;6;РТ;16;Йола;23
-Фамилия7;Имя7;ВУЗ1;РТФ;4;РТ;17;Йола;19
-Фамилия8;Имя8;ВУЗ1;РТФ;5;РТ;18;Йола;20
-Фамилия9;Имя9;ВУЗ1;РТФ;4;РТ;19;Йола;26
+﻿Фамилия1;Имя1;ВУЗ1;РТФ;РТ;5;11;Йола;18
+Фамилия2;Имя2;ВУЗ1;РТФ;РТ;6;12;Йола;19
+Фамилия3;Имя3;ВУЗ1;РТФ;РТ;5;13;Йола;20
+Фамилия8;Имя8;ВУЗ1;РТФ;РТ;5;18;Йола;20
+Фамилия4;Имя4;ВУЗ1;РТФ;РТ;6;14;Йола;21
+Фамилия5;Имя5;ВУЗ1;РТФ;РТ;6;15;Йола;22
+Фамилия6;Имя6;ВУЗ1;РТФ;РТ;6;16;Йола;23
+Фамилия7;Имя7;ВУЗ1;РТФ;РТ;4;17;Йола;19
+Фамилия9;Имя9;ВУЗ1;РТФ;РТ;4;19;Йола;26
